Fixed everything, cleaned up code
Removed components folder
Added src/components
Fixed index file
Fixed issue where graph would crash page
