--- Home.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-function Home(props) {
-    return (
-        <div>home component</div>
-    )
-}
-
-export default Home
