--- components/Enviromental Impact/index.js
@@ -1,2 +1,0 @@
-import EnviromentalImpact from "./Enviromental Impact"
-export default EnviromentalImpact
