--- components/Fuel Consumption/index.js
@@ -1,2 +1,0 @@
-import FuelConsumption from "./Fuel Consumption"
-export default FuelConsumption
