--- index.js
@@ -1,2 +1,0 @@
-import Home from "./Home"
-export default Home
