+++ src/components/App/Home.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function Home(props) {
+    return (
+        <div>home component</div>
+    )
+}
+
+export default Home
