+++ src/components/EnviromentalImpact/index.js
@@ -1,0 +1,2 @@
+import EnviromentalImpact from "./EnviromentalImpact"
+export default EnviromentalImpact
