+++ src/components/FuelConsumption/index.js
@@ -1,0 +1,2 @@
+import FuelConsumption from "./FuelConsumption"
+export default FuelConsumption
